DISCORD_XO/requirements.txt
fastapi>=0.100.0
uvicorn>=0.17.0
aiofiles>=0.8.0
discord.py>=2.0.0
aiosqlite>=0.17.0